refactored to a toon command with multiple subcommands to encompass functions
